implement string permutations function
`permuteString` is a function that recursively generates all
permutations of a string `str`.

It does this by taking the string, and starting at a given index
`currentIndex` (default 0), looping through the length of string,
swapping `currentIndex` for each of the other letters at greater
indices. At each iteration, it passes the new string back through
the function with `currentIndex` incremented to choose a new character
to swap with all of the further characters. When the recursive calls
have caused `currentIndex` to reach the end of the string, the base
case has been reached, and that string is added to the set of all
possible permutations.

This function is using a Set to prevent duplicate permutations for
strings with any repeated characters.

An additional test has been add for `permuteString` (made from the
double-checked output of an actual run).

Another test for `swapCharsAtIndices` and a slight modification for
the same has also been added to explicitly account for the case where
the indices of the characters to swap are equal.